DIP_LAB_2_BT23CSE182/pyproject.toml
[project]
name = "dip-lab-2-bt23cse182"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.31.0",
    "langchain-google-genai>=2.1.9",
    "loguru>=0.7.3",
    "matplotlib>=3.10.5",
    "numpy>=2.3.2",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
]
